Color switching
WIP color switching. colors are done as they are going to be until an ICP communication process begins. I need to know the terminal's current $TERMCOLOR but that's not possible until after the terminal's already launched. so I defaulted to xterm-256 for now but true color will return.
added 4 bit ansi color configurations.