breaking changes
broke ANSI colors on purpose. this is so we don't have a too many characters.
removed ot.ansi.color mostly
optimized the hack for enableing colors on conhost.exe and other windows terminals that also can be hacked. I use echo now with empty strings on the same line
the colors on terminal process will be fixed in the IPC updates to come but for now only trust the Host(main program no CLI UI)